% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_distribution_plot.R
\name{aa_distribution_plot}
\alias{aa_distribution_plot}
\title{Plotting of Amino Acid Distribution}
\usage{
aa_distribution_plot(aa_seq)
}
\arguments{
\item{aa_seq}{Sequence of amino acids}
}
\value{
A bar chart showing the counts of the different amino acids in the sequence.
}
\description{
Plotting of Amino Acid Distribution
}
\examples{
aa_distribution_plot('AACWIGI')
}
