% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_translation.R
\name{aa_translation}
\alias{aa_translation}
\title{Codon translation to aminoacids}
\usage{
aa_translation(codons)
}
\arguments{
\item{codons}{A, G, C and U written as a triplet}
}
\value{
returns a aminoacid string translated from codons
}
\description{
Codon translation to aminoacids
}
\examples{
aa_translation('CUA')
}
