% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_codons.R
\name{all_codons}
\alias{all_codons}
\title{Title : All codons in the given sequence}
\usage{
all_codons(sequence, start = 1)
}
\arguments{
\item{sequence}{a string representing a DNA or RNA sequence}

\item{start}{starting position to define where the codon extraction begins}
}
\value{
The function returns all the codons found in the provided DNA/RNA sequence.
}
\description{
Title : All codons in the given sequence
}
\examples{
all_codons("ATGGCT")
}
