% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_counts.R
\name{character_counts1}
\alias{character_counts1}
\title{Plot Character Counts in a Sequence}
\usage{
character_counts1(character_counts2)
}
\arguments{
\item{character_counts2}{A character string for which unique character counts will be calculated and plotted.}
}
\value{
A ggplot object showing a bar plot of counts of each unique character in the input string.
}
\description{
This function takes a character string input, splits it into unique characters, counts the occurrences of each character,
and returns a bar plot showing the counts for each unique character.
}
\examples{
# Example with a simple sequence
character_counts1("AABBCCDD")

# Example with a string containing amino acids
character_counts1("AAABBBCCC")
}
