% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_list.R
\name{codon_list}
\alias{codon_list}
\title{Title codon_list}
\usage{
codon_list(sequence, start = 1)
}
\arguments{
\item{sequence}{The name of the string that contains your amono acid sequence}

\item{start}{Set your reading frame (1, 2 or 3) (https://en.wikipedia.org/wiki/Reading_frame)}
}
\value{
Returns each codon from the sequence
}
\description{
Title codon_list
}
\examples{
codon_list(my_sequence, start = 2)
}
