% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codons_extraction.R
\name{codons_extraction}
\alias{codons_extraction}
\title{Codons extraction}
\usage{
codons_extraction(sequence, start = 1)
}
\arguments{
\item{sequence}{description}

\item{start}{of parameter}
}
\value{
description of function output
}
\description{
Codons extraction
}
