% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sequence.R
\name{concat_codons}
\alias{concat_codons}
\title{Concatenate Codons into a Single String}
\usage{
concat_codons(codons)
}
\arguments{
\item{codons}{A vector of indices corresponding to elements in `codon_table` to be concatenated.}
}
\value{
A single concatenated string formed by the elements at specified indices in `codon_table`.
}
\description{
Concatenate Codons into a Single String
}
