% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_DNA.R
\name{create_DNA}
\alias{create_DNA}
\title{DNA string creator}
\usage{
create_DNA(length)
}
\arguments{
\item{length}{The length (number of bases in the DNA string)}
}
\value{
A DNA string consisting of bases in string-format of length "length"
}
\description{
DNA string creator
}
\examples{
create_DNA(300)
}
