% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_sampler.R
\name{dna_sampler}
\alias{dna_sampler}
\title{dna_sampler}
\usage{
dna_sampler(sample_size)
}
\arguments{
\item{sample_size}{}
}
\value{
Returns a random sample of DNA, of specified size.
}
\description{
dna_sampler
}
\examples{
dna_sampler(10)
}
