% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_to_RNA.R
\name{DNA_to_RNA}
\alias{DNA_to_RNA}
\title{DNA to RNA}
\usage{
DNA_to_RNA(DNA_string)
}
\arguments{
\item{DNA_string}{The input to the function is a DNA string.
The function replaces T with U during the transcription.}
}
\value{
The function returns a RNA string transcribed from the given DNA string
}
\description{
DNA to RNA
}
\examples{
DNA_to_RNA("TACGT")
}
