% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequence_Generator.R
\name{generate_dna_sequence}
\alias{generate_dna_sequence}
\title{Random DNA Sequence Generator}
\usage{
generate_dna_sequence(length)
}
\arguments{
\item{length}{the length of the DNA sequence}
}
\value{
a random sequence of DNA as a string
}
\description{
Random DNA Sequence Generator
}
\examples{
ACGTTTCGAGT
}
