% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurence_plotter.R
\name{occurence_plotter}
\alias{occurence_plotter}
\title{Plotting the occurence of each amino acid}
\usage{
occurence_plotter(amino_acid_sequence)
}
\description{
Plotting the occurence of each amino acid
}
