% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_seq.R
\name{random_seq}
\alias{random_seq}
\title{Sampling a random sequence of a defined length}
\usage{
random_seq(len)
}
\arguments{
\item{len}{Length of the sequence to be sampled}
}
\value{
Sampled sequence
}
\description{
Sampling a random sequence of a defined length
}
\examples{
dna_seq <- random_seq(7)
}
