% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_to_U.R
\name{T_to_U}
\alias{T_to_U}
\title{T_to_U function
returns data with T bases transformed to U bases}
\usage{
T_to_U(data)
}
\arguments{
\item{data}{original data}
}
\description{
T_to_U function
returns data with T bases transformed to U bases
}
