% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transcription_table}
\alias{transcription_table}
\title{Title: transcription_table}
\format{

}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
transcription_table
}
\description{

Description: Tanslates RNA to amino acid.
}
\keyword{datasets}
