% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{translation}
\alias{translation}
\title{Translation}
\usage{
translation(codons)
}
\arguments{
\item{codons}{character vector of codons to translate}
}
\value{
chain of amino acids
}
\description{
Translation
}
\examples{
codon_table <- c("AUG" = "M", "UUU" = "F", "GGC" = "G")
translation(c("AUG","GGC","UUU"))
}
