% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{translation_table}
\alias{translation_table}
\title{Title: translation_table}
\format{

}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
translation_table
}
\description{
Description: Translates RNA to amino acid.
}
\keyword{datasets}
