% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_to_rna.R
\name{dna_to_rna}
\alias{dna_to_rna}
\title{Title: dna_to_rna}
\usage{
dna_to_rna(dna)
}
\arguments{
\item{dna}{sequence of DNA in a string}
}
\value{
rna sequence after T has been substituted by U
}
\description{
Title: dna_to_rna
}
\examples{
dna_seq <-"ATG"
rna_seq <- dna_to_rna(dna_seq)
}
